Giant Refactor
Deleted Debug.java. All it contained was a single method to print out a sequence of integers.

Created the FileIO and FileIOTests class. Other classes call methods of this class to perform file input/output operations.

Created the Main class. This class contains the main method, and allows the users to take quizzes until the quiz taking process fails.

Created the UserIO class. Other classes call methods of this class to perform input/output operations that interact with the user via console.

Created the Util class. Other classes call methods of this class to perform general, high level operations.

Other classes were modified in accordance with the newly created classes. For example, IO operations contained in the Quiz class are now contained in either the FileIO or UserIO classes.

Wrote the minimal requirements section in the software development journal. I wanted to put my software in a runnable state to make sure I did not forget anything. I have not really gotten there yet, but this commit is a step in the right direction.