Tokens, Arrays, Final
Implemented the token class and integrated it with the rest of the code. Each token consists of a lexeme, column number, and line number. This makes implementing parsing exceptions possible.

I realized that I was using array lists for everything, even when using collections of data of a fixed length. So, I switched to arrays where possible, and made a few methods that makes using them a little more convenient.

The classes Quiz, Question, and Token were full of getters and setters. They are really just used for structuring and reading data. So, I got rid of all the getters and setters in Question and Token, made their instance variables public, and made the classes immutable. I made Quiz immutable as well, but I had to keep one getter to return a question given an index.